Gson-Exception:  Added LoggingUtils and refactored the code
